♻️: created XMLService to handle fetching
